SCHED-171: Bugfix in instrument code and elimination of circular dependencies.
